BIG-24060: Set the price of the stencil master theme to 0.
In order for the stencil master theme to be set on newly created stores,
a "purchase" will need to be initiated. When the price of a theme is
greater than 0, we will require a valid stripe token.

This change will also allow existing merchants to "purchase" the master
theme without requiring a credit card.