Remove empty sidebar from category pages
https://github.com/bigcommerce/cornerstone/issues/1000

Three products are displayed per row on category pages if one or more of these three conditions are met.

A) faceted search is enabled.
B) Category contains subcategories.
C) shop by price is enabled

If none are met, the main content area remains at 75% width, leaving a quarter of the content area empty. This is due to the styling of .page-sidebar + .page-content.

Using the {{or}} operator, verify if either of the three conditions are met. If neither are true, page content will stretch 100% and display four products by default.

changelog