STENCIL-3533 Remove gift certificate format validation
Before we had a Gift Certificates API for the creation or import of gift certificates from another system, gift certificates in BC were all of a single format, alphanumeric with hyphens like this:

AAA-AAA-AAA-AAA

However, now we have the API and 3rd parties are welcome to import certificates of an arbitrary format.

However, Cornerstone still validates the old format, which will cause these imported certificates to fail.Making this change to fix that and allow any format.