Add global region at the bottom of the navigation
Currently, it's not possible to add a widget to a global region within the header. The only available region is "header_bottom--global" which is in the header template so is on all pages but is actually outside of the <header> element.

This region is useful for widgets that replace the navigation.