Load other font weights and styles for the body-font
Using a font name like `Google_Karla_400` for `body-font`, it would
only load the 400 font weight. When using `Google_Karla_700` for
another property, it would then load both the 400 and 700 fon weights.
But since in CSS you could make something bold or italic, the font
weight and/or style for this text might be missing, as it was not
loaded automatically by using it in another property.

This PR will make sure that the most common font weights and font
styles will be loaded for the body font.