feat(storefront): Dispatch an event on productOptionsChanged
Including all the information fetched in the previous request to the API to improve extensibility.

Having this mechanism, developers can set a listener and get as much information as they need regarding the selected option, so a new API call can be potentially avoided for further actions associated this event.